+++ CHANGELOG.md
@@ -3,6 +3,7 @@ In Devlopment
 - Serve a `/robots.txt` file denying all robots
 - Log `User-Agent` and `X-Request-ID` headers
 - Rate limit incoming requests
+- Handling of incoming requests now times out after 25 seconds
 
 v0.5.0 (2024-11-18)
 -------------------